deleted old preprocessing files
