+++ data/name_disambiguation/test_docs.csv
@@ -1,7 +1,6 @@
 ,au,au_org,au_person,cc,cc_org,collection,date,doc_type,pages,rc,rc_org,rc_person,text,tid,title,year
 0,,BATCo,"Dunn, WL; Garcia, Raquel",,,,,,,,,,,,,
 1,"Risi, Stephan","PM, PHILIP MORRIS",,,,,,,,,,,,,,
-2,"Risi, Stephan","BW, ATCO","Dunn, WL; Garcia, Raquel",,,,,,,,,,,,,
-2,"Risi, Stephan",xxxx,"Dunn, WL; Garcia, Raquel",,,,,,,,,,,,,
+1,,"BW, ATCO","Dunn, WL; Garcia, Raquel",,,,,,,,,,,,,
+1,"Risi, Stephan",xxxx,"Dunn, WL; Garcia, Raquel",,,,,,,,,,,,,
 1,"Risi, Stephan","PM, PHILIP MORRIS",,,,,,,,,,,,,,
-
